routes with permissions

+++ routes/api.php
@@ -41,105 +41,107 @@ Route::middleware('auth:api')->post('logout', [AuthController::class, 'logout'])
 Route::prefix('admin')
     ->middleware(['middleware' => 'auth:api'])
     ->group(function () {
-        Route::get('/dashboard', [DashboardController::class, 'index']);
+
+        // Dashboard
+        Route::get('/dashboard', [DashboardController::class, 'index'])
+            ->middleware('permission:dashboard-view');
+
+        // Activity Logs
         Route::get('/activity-logs', [ActivityLogController::class, 'index']);
         Route::post('/activity-logs/delete', [ActivityLogController::class, 'bulkDelete']);
+
         // Role and Permissions
         Route::controller(RolePermissionController::class)->group(function () {
-            Route::get('/roles', 'index');
-            Route::get('/roles/show/{id}', 'show');
-            Route::post('/roles/create', 'store');
-            Route::post('/roles/update/{id}', 'update');
-            Route::post('/permissions/add/{id}', 'addPermissions');
-            Route::post('/permissions/remove/{id}', 'removePermissions');
-            Route::delete('/roles/delete/{id}', 'destroy');
+            Route::get('/roles', 'index')->middleware('permission:role-list');
+            Route::get('/roles/show/{id}', 'show')->middleware('permission:role-manage');
+            Route::post('/roles/create', 'store')->middleware('permission:role-manage');
+            Route::post('/roles/update/{id}', 'update')->middleware('permission:role-manage');
+            Route::post('/permissions/add/{id}', 'addPermissions')->middleware('permission:role-manage');
+            Route::post('/permissions/remove/{id}', 'removePermissions')->middleware('permission:role-manage');
+            Route::delete('/roles/delete/{id}', 'destroy')->middleware('permission:role-manage');
             Route::get('/permissions', 'getPermissions');
         });
 
         // User Management
         Route::controller(UserManagementController::class)->group(function () {
-            Route::get('/users', 'index');
-            Route::get('/users/show/{id}', 'show');
-            Route::post('/users/create', 'store');
-            Route::post('/users/update/{id}', 'update');
-            Route::patch('/users/toggle-approval/{id}', 'toggleApproval');
-            Route::delete('/users/delete/{id}', 'destroy');
+            Route::get('/users', 'index')->middleware('permission:user-list');
+            Route::get('/users/show/{id}', 'show')->middleware('permission:user-manage');
+            Route::post('/users/create', 'store')->middleware('permission:user-manage');
+            Route::post('/users/update/{id}', 'update')->middleware('permission:user-manage');
+            Route::patch('/users/toggle-approval/{id}', 'toggleApproval')->middleware('permission:user-manage');
+            Route::delete('/users/delete/{id}', 'destroy')->middleware('permission:user-manage');
             Route::get('/users/roles', 'getRoles');
         });
 
-        //Blogs Management
+        // Blogs Management
         Route::controller(BlogsController::class)->prefix('blogs')->group(function () {
-            Route::get('/', 'index');
-            Route::get('/show/{id}', 'show');
-            Route::post('/create', 'store');
-            Route::post('/update/{id}', 'update');
-            Route::delete('/delete/{id}', 'destroy');
+            Route::get('/', 'index')->middleware('permission:blog-list');
+            Route::get('/show/{id}', 'show')->middleware('permission:blog-list');
+            Route::post('/create', 'store')->middleware('permission:blog-manage');
+            Route::post('/update/{id}', 'update')->middleware('permission:blog-manage');
+            Route::delete('/delete/{id}', 'destroy')->middleware('permission:blog-manage');
         });
 
-        //Testimonials Management
+        // Testimonials Management
         Route::controller(TestimonialsController::class)->prefix('testimonials')->group(function () {
-            Route::get('/', 'index');
-            Route::get('/show/{id}', 'show');
-            Route::post('/create', 'store');
-            Route::post('/update/{id}', 'update');
-            Route::delete('/delete/{id}', 'destroy');
-            Route::post('/toggle-status/{id}', 'toggleStatus');
+            Route::get('/', 'index')->middleware('permission:testimonial-list');
+            Route::get('/show/{id}', 'show')->middleware('permission:testimonial-list');
+            Route::post('/create', 'store')->middleware('permission:testimonial-manage');
+            Route::post('/update/{id}', 'update')->middleware('permission:testimonial-manage');
+            Route::delete('/delete/{id}', 'destroy')->middleware('permission:testimonial-manage');
+            Route::post('/toggle-status/{id}', 'toggleStatus')->middleware('permission:testimonial-manage');
         });
 
-        //Contact Inquiries
+        // Contact Inquiries
         Route::controller(ContactSubmissionController::class)->group(function () {
-            Route::get('/contact-enquiries', 'index');
-            Route::get('/contact-enquiries/show/{id}', 'show');
-            Route::delete('/contact-enquiries/delete/{id}', 'destroy');
+            Route::get('/contact-enquiries', 'index')->middleware('permission:contact-inquiry-list');
+            Route::get('/contact-enquiries/show/{id}', 'show')->middleware('permission:contact-inquiry-view');
+            Route::delete('/contact-enquiries/delete/{id}', 'destroy')->middleware('permission:contact-inquiry-delete');
         });
 
-        //Inspection Inquiries
+        // Inspection Inquiries
         Route::controller(InspectionEnquiryController::class)->group(function () {
-            Route::get('/inspection-enquiries', 'index');
-            Route::get('/inspection-enquiries/show/{id}', 'show');
-            Route::delete('/inspection-enquiries/delete/{id}', 'destroy');
+            Route::get('/inspection-enquiries', 'index')->middleware('permission:inspection-list');
+            Route::get('/inspection-enquiries/show/{id}', 'show')->middleware('permission:inspection-view');
+            Route::delete('/inspection-enquiries/delete/{id}', 'destroy')->middleware('permission:inspection-delete');
         });
 
-        //Purchase Inquiries
+        // Purchase Inquiries
         Route::controller(PurchaseEnquiryController::class)->group(function () {
-            Route::get('/purchase-enquiries', 'index');
-            Route::get('/purchase-enquiries/show/{id}', 'show');
-            Route::delete('/purchase-enquiries/delete/{id}', 'destroy');
+            Route::get('/purchase-enquiries', 'index')->middleware('permission:purchase-inquiry-list');
+            Route::get('/purchase-enquiries/show/{id}', 'show')->middleware('permission:purchase-inquiry-view');
+            Route::delete('/purchase-enquiries/delete/{id}', 'destroy')->middleware('permission:purchase-inquiry-delete');
         });
 
-        //Sale Inquiries
+        // Sale Inquiries
         Route::controller(SaleEnquiryController::class)->group(function () {
-            Route::get('/sale-enquiries', 'index');
-            Route::get('/sale-enquiries/show/{id}', 'show');
-            Route::delete('/sale-enquiries/delete/{id}', 'destroy');
+            Route::get('/sale-enquiries', 'index')->middleware('permission:sale-inquiry-list');
+            Route::get('/sale-enquiries/show/{id}', 'show')->middleware('permission:sale-inquiry-view');
+            Route::delete('/sale-enquiries/delete/{id}', 'destroy')->middleware('permission:sale-inquiry-delete');
         });
 
-        //Bid Management
+        // Bid Management
         Route::controller(BidManagementController::class)->group(function () {
-            Route::get('/bids', 'index');
-            Route::get('/bids/show/{id}', 'show');
-            Route::patch('/bids/toggle-status/{id}', 'toggleStatus');
-            Route::patch('/bids/reject/{id}', 'reject');
-            Route::delete('/bids/delete/{id}', 'destroy');
-            Route::post('/bids/bulk-delete', 'bulkDelete');
+            Route::get('/bids', 'index')->middleware('permission:bidding-list');
+            Route::get('/bids/show/{id}', 'show')->middleware('permission:bidding-list');
+            Route::patch('/bids/toggle-status/{id}', 'toggleStatus')->middleware('permission:bidding-actions');
+            Route::patch('/bids/reject/{id}', 'reject')->middleware('permission:bidding-actions');
+            Route::delete('/bids/delete/{id}', 'destroy')->middleware('permission:bidding-actions');
+            Route::post('/bids/bulk-delete', 'bulkDelete')->middleware('permission:bidding-actions');
         });
 
-        //Makes Management
+        // Makes Management
         Route::controller(MakeController::class)->group(function () {
-            Route::get('/makes', 'index');
-            Route::get('/makes/show/{id}', 'show');
-            Route::post('/makes/create', 'store');
-            Route::post('/makes/update/{id}', 'update');
-            Route::delete('/makes/delete/{id}', 'destroy');
-            Route::delete('/model/delete/{id}', 'deleteModel');
-        });
-
-        //Model By Make
-        Route::controller(MakeController::class)->group(function () {
+            Route::get('/makes', 'index')->middleware('permission:make-list');
+            Route::get('/makes/show/{id}', 'show')->middleware('permission:make-list');
+            Route::post('/makes/create', 'store')->middleware('permission:make-actions');
+            Route::post('/makes/update/{id}', 'update')->middleware('permission:make-actions');
+            Route::delete('/makes/delete/{id}', 'destroy')->middleware('permission:make-actions');
+            Route::delete('/model/delete/{id}', 'deleteModel')->middleware('permission:make-actions');
             Route::get('/make-models/{id}', 'modelsByMake');
         });
 
-        //Vehicles Management
+        // Vehicles Management
         Route::prefix('vehicles')->controller(VehicleManagementController::class)->group(function () {
             Route::get('body-types', 'getBodyTypes');
             Route::get('fuel-types', 'getFuelTypes');
@@ -148,44 +150,38 @@ Route::prefix('admin')
             Route::get('features/exterior', 'getExteriorFeatures');
             Route::get('features/interior', 'getInteriorFeatures');
             Route::get('tags', 'getTags');
-            Route::get('/', 'index');
-            Route::get('show/{id}', 'show');
-            Route::post('create', 'store');
-            Route::post('update/{id}', 'update');
+            Route::get('/', 'index')->middleware('permission:vehicle-list');
+            Route::get('show/{id}', 'show')->middleware('permission:vehicle-view');
+            Route::post('create', 'store')->middleware('permission:vehicle-create');
+            Route::post('update/{id}', 'update')->middleware('permission:vehicle-edit');
             Route::post('images/add/{id}', 'addImages');
             Route::post('images/remove/{id}', 'removeImages');
-            Route::delete('delete/{id}', 'destroy');
+            Route::delete('delete/{id}', 'destroy')->middleware('permission:vehicle-delete');
         });
         Route::controller(VehicleManagementController::class)->group(function () {
-          
             Route::get('/auctions', 'auctions');
-           
         });
 
-        //Inspection Report Management
+        // Inspection Report Management
         Route::controller(InspectionReportController::class)->group(function () {
-            Route::get('/inspection-reports', 'index');
-            Route::get('/inspection-reports/show/{id}', 'show');
-            Route::post('/inspection-reports/create', 'store');
-
-            Route::post('/inspection-reports/upload-images', 'storeVehicleImages');
-            Route::post('/inspection-reports/remove-images', 'removeVehicleImages');
-
-            Route::post('/inspection-reports/field-images/add', 'storeInspectionFields');
-            Route::post('/inspection-reports/field-images/remove', 'removeInspectionFieldImages');
-
+            Route::get('/inspection-reports', 'index')->middleware('permission:inspection-list');
+            Route::get('/inspection-reports/show/{id}', 'show')->middleware('permission:report-view');
+            Route::post('/inspection-reports/create', 'store')->middleware('permission:report-create');
+            Route::post('/inspection-reports/upload-images', 'storeVehicleImages')->middleware('permission:report-edit');
+            Route::post('/inspection-reports/remove-images', 'removeVehicleImages')->middleware('permission:report-edit');
+            Route::post('/inspection-reports/field-images/add', 'storeInspectionFields')->middleware('permission:report-edit');
+            Route::post('/inspection-reports/field-images/remove', 'removeInspectionFieldImages')->middleware('permission:report-edit');
             Route::get('/inspection-reports/damage-types', 'getDamageTypes');
-            Route::post('/inspection-reports/damage/add', 'addDamage');
-            Route::post('/inspection-reports/damage/remove', 'removeDamage');
-
-            Route::post('/inspection-reports/update/{id}', 'update');
-            Route::delete('/inspection-reports/delete/{id}', 'destroy');
-            Route::post('/inspection-reports/generate-pdf/{id}', 'generatePdf');
-            Route::post('/inspection-reports/share-link/{id}', 'share');
-            Route::get('/inspection-reports/download/{id}', 'downloadReport');
+            Route::post('/inspection-reports/damage/add', 'addDamage')->middleware('permission:report-edit');
+            Route::post('/inspection-reports/damage/remove', 'removeDamage')->middleware('permission:report-edit');
+            Route::post('/inspection-reports/update/{id}', 'update')->middleware('permission:report-edit');
+            Route::delete('/inspection-reports/delete/{id}', 'destroy')->middleware('permission:report-delete');
+            Route::post('/inspection-reports/generate-pdf/{id}', 'generatePdf')->middleware('permission:report-generate-pdf');
+            Route::post('/inspection-reports/share-link/{id}', 'share')->middleware('permission:report-share');
+            Route::get('/inspection-reports/download/{id}', 'downloadReport')->middleware('permission:report-download');
         });
 
-        //Notifications
+        // Notifications
         Route::controller(NotificationController::class)->group(function () {
             Route::get('/notifications', 'index');
             Route::get('/notifications/get-all', 'allNotifications');
@@ -196,7 +192,7 @@ Route::prefix('admin')
             Route::delete('/notifications/clear', 'clearAll');
         });
 
-
+        // Agents
         Route::controller(AgentManagementController::class)->group(function () {
             Route::get('/agents', 'index');
             Route::post('/agents/create', 'store');
@@ -206,8 +202,9 @@ Route::prefix('admin')
             Route::get('/unassigned-customers/{id}', 'customersByAgent');
             Route::delete('/agents/delete/{id}', 'destroy');
         });
-        //Packges
-         Route::controller(PackageController::class)->group(function () {
+
+        // Packages
+        Route::controller(PackageController::class)->group(function () {
             Route::get('/packages', 'index');
             Route::post('/packages/create', 'store');
             Route::get('/packages/show/{id}', 'show');
@@ -215,13 +212,11 @@ Route::prefix('admin')
             Route::delete('/packages/delete/{id}', 'destroy');
         });
 
-        //Profile
+        // Profile
         Route::controller(ProfileController::class)->group(function () {
             Route::get('/profile', 'index');
         });
     });
-
-
